backend_chamazetu/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
azure-core==1.30.1
azure-identity==1.16.0
azure-storage-blob==12.19.1
bcrypt==4.1.2
boto3==1.34.118
botocore==1.34.118
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.4
deprecation==2.1.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.0
fastapi-jwt-auth==0.5.0
gotrue==2.1.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
jose==1.0.0
Mako==1.3.0
MarkupSafe==2.1.3
msal==1.28.0
msal-extensions==1.1.0
orjson==3.9.10
packaging==23.2
passlib==1.7.4
pillow==10.3.0
portalocker==2.8.2
postgrest==0.13.2
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic-extra-types==2.4.1
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyJWT==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
realtime==1.0.2
requests==2.31.0
rsa==4.9
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.3
supafunc==0.3.3
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==11.0.3
